// src/types/generated/RefreshPlan.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RefreshSourceDecision } from "./RefreshSourceDecision";

export type RefreshPlan = { decisions: Array<RefreshSourceDecision>, refresh_count: number, champ_select_blocked: boolean, };