// src/types/generated/FetchLogSummary.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SourceFetchHealth } from "./SourceFetchHealth";

export type FetchLogSummary = { sources: Array<SourceFetchHealth>, total_entries: number, healthy_count: number, degraded_count: number, };